+++ src/components/preview/RepeatedThought.vue
@@ -1,13 +1,9 @@
 <template>
-  <div>
-    <h5 style="margin-bottom: 5px;">{{ data.title }}</h5>
-    <span v-html="data.text"/>
-  </div>
+  <div>My component</div>
 </template>
 
 <script>
 export default {
-  props: ['data'],
   // name: 'ComponentName',
   data () {
     return {}